+++ .nvmrc
@@ -1,1 +1,5 @@
+<<<<<<< Updated upstream
 10.10.0
+=======
+10.11.0
+>>>>>>> Stashed changes
