+++ src/factories/Message/index.js
@@ -1,48 +1,13 @@
+const generateMessage = require('./generate');
+const formatMessage = require('./format');
+
 const Message = function Message(type = 'slack', team = {}) {
-  const {
-    color = '#3ef2c5', mentions = '', emoji = '🌱', teamMessages = '', name = 'General', teamTitles = '',
-  } = team;
-  const teamName = `${name} ${emoji}`;
-
-  const generateSlack = function generateSlack(message = '') {
-    if (!teamMessages && !message) return '';
-
-    let fields = [];
-    if (teamMessages) {
-      fields = [
-        {
-          title: 'Stories',
-          value: teamMessages,
-          short: true,
-        },
-        {
-          title: 'Details',
-          value: teamTitles,
-          short: true,
-        },
-      ];
-    }
-
-    return {
-      fallback: teamMessages || message,
-      color,
-      title: teamName,
-      text: mentions,
-      fields,
-    };
-  };
-
-  const generate = function generate(message = '') {
-    switch (type) {
-      case 'slack':
-        return generateSlack(message);
-      default:
-        return '';
-    }
-  };
+  const generate = generateMessage(type, team);
+  const format = formatMessage(team);
 
   return {
     generate,
+    format,
   };
 };
